Fix camera timeout.
